requirements.txt
Flask==3.0.3
flask-cors==4.0.1
werkzeug==3.0.3
torch==2.3.0
torchvision==0.18.0
numpy==1.26.4
matplotlib==3.8.4
opencv-python-headless==4.9.0.80
h5py==3.11.0
Pillow==10.3.0
gunicorn==22.0.0
